Start the editing-a-return journey from the licence Returns tab and use role-based locators throughout
Selects the return from its table row on the Returns tab rather than jumping straight to the return details URL, and replaces .govuk-button CSS locators with accessible getByRole button/link locators plus h1 assertions for each page in the journey.
